pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = "setuptools.build_meta"

[project]
name = "mllibprodest"
version = "1.5.0"
license = {text = "GPLv3"}
authors = [
  { name="Instituto de Tecnologia da Informação e Comunicação do Espírito Santo (PRODEST)", email="prodest@prodest.es.gov.br" },
]
description = "Biblioteca de Machine Learning (ML) do Prodest."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: Portuguese (Brazilian)"
]
dependencies = ['minio==7.1.12', 'python-dotenv==0.21.0', 'mlflow==1.29.0', 'boto3==1.24.89']
keywords = ['Prodest', 'ML', 'lib', 'stack']

[project.urls]
"Homepage" = "https://github.com/prodest/mllibprodest"
"Bug Tracker" = "https://github.com/prodest/mllibprodest/issues"
"Documentation" = "https://github.com/prodest/mllibprodest/blob/main/README.md"
